requirements.txt
# Core libraries
praw              # For accessing Reddit API
kafka-python      # For producing/consuming Kafka messages
influxdb-client   # For writing/reading data to InfluxDB 2.x